+++ packages/thumbprint-react/CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Changed
 
 -   [Patch] Use newer version of TypeScript as dev dependency.
+-   [Patch] Mark the `shouldOpenInNewTab` prop of `Link` and `ThemedLink` as deprecated.
+
+### Added
+
+-   [Minor] Add `target` prop to `Link` and `ThemedLink`.
 
 ## 14.0.0 - 2020-07-14